eureka-server -> eureka_service rename
