+++ .github/workflows/build_and_deploy.yml
@@ -38,7 +38,7 @@ jobs:
       - name: Eureka Build and Push to Dockerhub
         uses: docker/build-push-action@v2
         with:
-          file: /eureka_server/Dockerfile
+          file: /eureka_service/Dockerfile
           push: true
           tags: nekromant322/overmoney:eureka
