+++ eureka_service/Dockerfile
@@ -1,8 +1,8 @@
 FROM maven:3.6.3-jdk-11-slim AS build
 RUN mkdir -p /eureka
 WORKDIR /eureka
-COPY /eureka_server/pom.xml /eureka
-COPY /eureka_server/src /eureka/src
+COPY /eureka_service/pom.xml /eureka
+COPY /eureka_service/src /eureka/src
 RUN mvn -B -f pom.xml clean package -DskipTests
 
 FROM openjdk:11-jdk-slim
