+++ eureka_service/pom.xml
@@ -9,10 +9,10 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <groupId>com.override</groupId>
-    <artifactId>eureka_server</artifactId>
+    <artifactId>eureka_service</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>eureka_server</name>
-    <description>eureka_server</description>
+    <name>eureka_service</name>
+    <description>eureka_service</description>
     <properties>
         <java.version>11</java.version>
         <spring-cloud.version>2022.0.2</spring-cloud.version>
