+++ eureka_service/src/main/java/com/override/eureka_service/EurekaServiceApplication.java
@@ -1,4 +1,4 @@
-package com.override.eureka_server;
+package com.override.eureka_service;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
 @SpringBootApplication
 @EnableEurekaServer
-public class EurekaServerApplication {
+public class EurekaServiceApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(EurekaServerApplication.class, args);
+        SpringApplication.run(EurekaServiceApplication.class, args);
     }
 
 }
