+++ eureka_service/src/main/resources/application.yml
@@ -1,6 +1,6 @@
   spring:
     application:
-      name: eureka_server
+      name: eureka_service
 
   server:
     port: 8761
