+++ eureka_service/src/test/java/com/override/eureka_service/EurekaServiceApplicationTests.java
@@ -1,10 +1,9 @@
-package com.override.eureka_server;
+package com.override.eureka_service;
 
-import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class EurekaServerApplicationTests {
+class EurekaServiceApplicationTests {
 
 //    @Test
 //    void contextLoads() {
