+++ pom.xml
@@ -16,7 +16,7 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>eureka_server</module>
+        <module>eureka_service</module>
         <module>recognizer_service</module>
         <module>orchestrator_service</module>
         <module>telegram_bot_service</module>
